modificacion migraciones

+++ database/migrations/2024_01_16_160316_create_yagravantes_table.php
@@ -16,21 +16,21 @@ return new class extends Migration
         Schema::create('yagravantes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('inspeccion_id')->constrained();
-            $table->integer('requiere_extintor')->nullable();
-            $table->integer('cuenta_extintor')->nullable();
-            $table->integer('extintores_apropiados')->nullable();
-            $table->integer('distribuidos')->nullable();
-            $table->integer('señalizados')->nullable();
-            $table->integer('despejados')->nullable();
-            $table->integer('instalados')->nullable();
-            $table->integer('recargados')->nullable();
-            $table->integer('buenos_extintores')->nullable();
-            $table->integer('supervision')->nullable();
-            $table->integer('aplica_seguridad_h')->nullable();
-            $table->integer('aplica_planes')->nullable();
-            $table->integer('seguro_vigente')->nullable();
-            $table->integer('poliza_cubre')->nullable();
-            $table->integer('compania_seguro')->nullable();
+            $table->tinyInteger('requiere_extintor')->nullable();
+            $table->tinyInteger('cuenta_extintor')->nullable();
+            $table->tinyInteger('extintores_apropiados')->nullable();
+            $table->tinyInteger('distribuidos')->nullable();
+            $table->tinyInteger('señalizados')->nullable();
+            $table->tinyInteger('despejados')->nullable();
+            $table->tinyInteger('instalados')->nullable();
+            $table->tinyInteger('recargados')->nullable();
+            $table->tinyInteger('buenos_extintores')->nullable();
+            $table->tinyInteger('supervision')->nullable();
+            $table->tinyInteger('aplica_seguridad_h')->nullable();
+            $table->tinyInteger('aplica_planes')->nullable();
+            $table->tinyInteger('seguro_vigente')->nullable();
+            $table->tinyInteger('poliza_cubre')->nullable();
+            $table->tinyInteger('compania_seguro')->nullable();
 
             $table->timestamps();
         });
